fix: resolve Windows CI failures from unused variables
- Move TempDir import to Unix-only test block where it's used
- Remove unnecessary usize cast on Windows (WorkingSetSize already usize)
- Add Windows cross-check to justfile lint command for early detection

+++ crates/maos/src/io/processor.rs
@@ -297,7 +297,7 @@ impl StdinProcessor {
                     std::mem::size_of::<PROCESS_MEMORY_COUNTERS>() as u32,
                 );
                 if result != 0 {
-                    return Some(counters.WorkingSetSize as usize);
+                    return Some(counters.WorkingSetSize);
                 }
             }
         }
